Move get_scilista_file_path p/ operations
Assim, essa função pode ser usada tanto pela DAG de pré sync existente 
quanto pela nova DAG de coleta de pacotes.
