+++ airflow/dags/operations/pre_sync_documents_to_kernel_operations.py
@@ -10,6 +10,33 @@ PREFIX_PATTERN += ("-" + ("[0-9]" * 2)) * 5
 PREFIX_PATTERN += "-" + ("[0-9]" * 6)
 
 
+def get_scilista_file_path(
+    xc_sps_packages_dir: Path, proc_sps_packages_dir: Path, gerapadrao_id: str
+) -> str:
+    """Garante que a scilista usada será a do diretório de PROC. Quando for a primeira
+    execução da DAG, a lista será copiada para o diretório de PROC. Caso contrário, a 
+    mesma será mantida.
+    """
+    proc_dir_scilista_list = list(proc_sps_packages_dir.glob(f"scilista-*.lst"))
+    if proc_dir_scilista_list:
+        _proc_scilista_file_path = proc_dir_scilista_list[0]
+        Logger.info('Proc scilista "%s" already exists', _proc_scilista_file_path)
+    else:
+        _scilista_filename = f"scilista-{gerapadrao_id}.lst"
+        _origin_scilista_file_path = xc_sps_packages_dir / _scilista_filename
+        if not _origin_scilista_file_path.is_file():
+            raise FileNotFoundError(_origin_scilista_file_path)
+
+        _proc_scilista_file_path = proc_sps_packages_dir / _scilista_filename
+        Logger.info(
+            'Copying original scilista "%s" to proc "%s"',
+            _origin_scilista_file_path,
+            _proc_scilista_file_path,
+        )
+        shutil.copy(_origin_scilista_file_path, _proc_scilista_file_path)
+    return str(_proc_scilista_file_path)
+
+
 def get_sps_packages(scilista_file_path, xc_dir_name, proc_dir_name):
     """
     Obtém Pacotes SPS através da Scilista, movendo os pacotes para o diretório de
