+++ airflow/dags/pre_sync_documents_to_kernel.py
@@ -39,33 +39,6 @@ dag = DAG(
 )
 
 
-def get_scilista_file_path(
-    xc_sps_packages_dir: Path, proc_sps_packages_dir: Path, gerapadrao_id: str
-) -> str:
-    """Garante que a scilista usada será a do diretório de PROC. Quando for a primeira
-    execução da DAG, a lista será copiada para o diretório de PROC. Caso contrário, a 
-    mesma será mantida.
-    """
-    proc_dir_scilista_list = list(proc_sps_packages_dir.glob(f"scilista-*.lst"))
-    if proc_dir_scilista_list:
-        _proc_scilista_file_path = proc_dir_scilista_list[0]
-        Logger.info('Proc scilista "%s" already exists', _proc_scilista_file_path)
-    else:
-        _scilista_filename = f"scilista-{gerapadrao_id}.lst"
-        _origin_scilista_file_path = xc_sps_packages_dir / _scilista_filename
-        if not _origin_scilista_file_path.is_file():
-            raise FileNotFoundError(_origin_scilista_file_path)
-
-        _proc_scilista_file_path = proc_sps_packages_dir / _scilista_filename
-        Logger.info(
-            'Copying original scilista "%s" to proc "%s"',
-            _origin_scilista_file_path,
-            _proc_scilista_file_path,
-        )
-        shutil.copy(_origin_scilista_file_path, _proc_scilista_file_path)
-    return str(_proc_scilista_file_path)
-
-
 def get_sps_packages(conf, **kwargs):
     """Executa ``pre_sync_documents_to_kernel_operations.get_sps_packages`` com a 
     scilista referente à DagRun. Todos os pacotes obtidos serão armazenados em 
@@ -79,7 +52,7 @@ def get_sps_packages(conf, **kwargs):
     if not _proc_sps_packages_dir.is_dir():
         _proc_sps_packages_dir.mkdir()
 
-    _scilista_file_path = get_scilista_file_path(
+    _scilista_file_path = pre_sync_documents_to_kernel_operations.get_scilista_file_path(
         _xc_sps_packages_dir,
         _proc_sps_packages_dir,
         Variable.get("GERAPADRAO_ID_FOR_SCILISTA"),
